requirements.txt
numpy>=1.24.2
wandb>=0.13.10
matplotlib>=3.7.0
scipy>=1.10.0
pytorch-lightning>=2.0.3
shapely>=2.0.1
networkx>=3.0
Cartopy>=0.22.0
pyproj>=3.4.1
tueplots>=0.0.8
plotly>=5.15.0
torch_geometric
xarray
zarr
s3fs
